requirements.txt
scipy
appnope==0.1.0
biopython==1.65
decorator==4.0.4
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
msgpack-python==0.4.6
networkx==1.10
numexpr==2.4.4
numpy==1.10.1
pandas==0.17.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.7
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wsgiref==0.1.2
simplejson==3.8.1
python-Levenshtein
